% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bootstrap_slr}
\alias{bootstrap_slr}
\title{Bootstrap Summary Table}
\usage{
bootstrap_slr(boot_slr, ols_slr)
}
\arguments{
\item{boot_slr}{Bootstrap coefficients generate by bootstrap_slr_summary()}

\item{ols_slr}{OLS's coefficients generate by ols_estimators()}
}
\value{
A data.frame with rows for intercept and slope and columns.
}
\description{
Calculates descriptive statistics for the selected
numerical variables, including mean, median, standard deviation,etc.
}
\examples{
data(BostonHousing)
data <- BostonHousing
respond <- "medv"
predictor <-"lstat"
respond <- "lstat"
predictor <-"medv"
ols_slr <- ols_estimators(data,respond, predictor)
boot_slr<-bootstrap_slr_summary(data, R = 1000,seed = NULL,predictor,respond)
boot_table<-bootstrap_slr(boot_slr,ols_slr)
}
