% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bootstrap_slr_summary}
\alias{bootstrap_slr_summary}
\title{Bootstrap analysis}
\usage{
bootstrap_slr_summary(data, R = 1000, seed = NULL, predictor, respond)
}
\arguments{
\item{data}{Boston Housoing dataset}

\item{R}{Number of bootstrap samples (set as 1000)}

\item{seed}{Input seed for reproducibility}

\item{predictor}{A string that contains the name of the predictor variable.}

\item{respond}{A string that contains the name of the response variable.}
}
\value{
A list containing: Bootstrap bias estimates,standard errors, etc
}
\description{
Bootstrap analysis
}
\examples{
data <- BostonHousing
respond <- "medv"
predictor <-"lstat"
ols_slr <- ols_estimators(data,respond, predictor)
boot_slr<-bootstrap_slr_summary(data, R = 1000,seed = NULL,predictor,respond)
}
