% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{BostonHousing}
\alias{BostonHousing}
\title{Boston Housing Dataset}
\format{
An object of class \code{data.frame} with 506 rows and 14 columns.
}
\usage{
BostonHousing
}
\description{
Boston Housing Dataset
}
\keyword{datasets}
