% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{interactive_single_slr}
\alias{interactive_single_slr}
\title{Interactive Single SLR + Bootstrap}
\usage{
interactive_single_slr(data, R = 1000, seed = 2025, verbose = TRUE)
}
\arguments{
\item{data}{Boston Housing data frame}

\item{R}{Bootstrap replications, set default as 1000}

\item{seed}{Random seed, set default as 2025}

\item{verbose}{Show messages}
}
\value{
Same as run_single_slr()
}
\description{
Interactively select variables and run analysis.
}
\examples{
data <- BostonHousing
if(interactive()) {
interactive_single_slr(data)
}
}
