% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ols_estimators}
\alias{ols_estimators}
\title{Generate simple linear regression model}
\usage{
ols_estimators(data, predictor, respond)
}
\arguments{
\item{data}{Boston Housing dataset.}

\item{predictor}{A string that contains the name of the predictor variable.}

\item{respond}{A string that contains the name of the response variable.}
}
\value{
A list containing: beta0, beta1, ols_conf_int
(95\% confidence interval).
}
\description{
Help to calculate the OLS estimators generate by using
Boston Housing dataset.
}
\examples{
data <- BostonHousing
respond <- "medv"
predictor <- "lstat"
ols_slr <- ols_estimators(data, predictor, respond)
print(ols_slr)
}
