% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_ci_box}
\alias{plot_ci_box}
\title{Plot confidence interval comparison between bootstrap and OLS}
\usage{
plot_ci_box(
  ci_list,
  term = c("intercept", "slope"),
  main = NULL,
  col_boots = "lightblue",
  col_ols = "red"
)
}
\arguments{
\item{ci_list}{List containing boot_perc and ols_wald matrices}

\item{term}{One of "intercept" or "slope"}

\item{main}{Optional title}

\item{col_ols}{Color of OLS ci graph. Set "red" as default.}

\item{col_hist}{Color of bootstrap graph. Set "lightblue" as default.}

\item{...}{Additional args passed to hist()
respond <- "medv"
predictor <-"lstat"
ols_slr <- ols_estimators(data, predictor, respond)
boot_slr<-bootstrap_slr_summary(data, R = 1000,seed = NULL,predictor,respond)
ci_list<- bootstrap_slr_ci(boot_slr, ols_slr, level = 0.95)
plot_ci_box(ci_list, "intercept")}
}
\value{
Invisibly returns the data used for plotting
}
\description{
Plot confidence interval comparison between bootstrap and OLS
}
