% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_iqr_boxplot}
\alias{plot_iqr_boxplot}
\title{Generate boxplot and calculate IQR for selected variables}
\usage{
plot_iqr_boxplot(
  data,
  predictor,
  respond,
  col_box = c("lightblue", "lightgreen"),
  main = NULL,
  ylab = "Value",
  ...
)
}
\arguments{
\item{data}{A data.frame containing the variables.}

\item{predictor}{A string that contains the name of the predictor variable.}

\item{respond}{A string that contains the name of the response variable.}

\item{col_box}{Colors of the boxplots. Default c("lightblue", "lightgreen").}

\item{main}{Optional title. If NULL, generates automatic title.}

\item{ylab}{Optional y-axis label.}

\item{...}{Additional args passed to boxplot()}
}
\value{
Invisibly returns the boxplot object.
}
\description{
Generate boxplot and calculate IQR for selected variables
}
\examples{
data <- BostonHousing
respond <- "medv"
predictor <- "lstat"
plot_iqr_boxplot(data, predictor, respond)
}
