% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_lr_bootstrap_scatter}
\alias{plot_lr_bootstrap_scatter}
\title{LR vs Bootstrap Scatter Comparison Plot}
\usage{
plot_lr_bootstrap_scatter(
  ols_slr,
  boot_slr,
  color_boot = "#56B4E9",
  color_contour = "#2C3E50",
  color_lr = "#E69F00",
  color_mean = "#009E73",
  color_arrow = "#E74C3C",
  alpha_boot = 0.5,
  alpha_contour = 0.5,
  size_boot = 2,
  size_lr = 6,
  size_mean = 5,
  size_arrow = 1
)
}
\arguments{
\item{ols_slr}{A list containing OLS estimates made by ols_estimators()}

\item{boot_slr}{A list containing bootstrap estimates made
by bootstrap_slr_summary()}

\item{color_boot}{Bootstrap points color}

\item{color_contour}{Density contour color}

\item{color_lr}{LR estimate point color}

\item{color_mean}{Bootstrap mean point color}

\item{color_arrow}{Bias arrow color}

\item{alpha_boot}{Transparency of bootstrap points}

\item{alpha_contour}{Transparency of density contours}

\item{size_boot}{Size of bootstrap points}

\item{size_lr}{Size of LR estimate point}

\item{size_mean}{Size of mean point}

\item{size_arrow}{arrow size}
}
\value{
A ggplot object of LR vs Bootstrap Scatter Comparison Plot
}
\description{
Creates a scatter plot comparing linear regression model's
estimators with bootstrap parameter estimates.
}
\examples{
data(BostonHousing)
data <- BostonHousing
respond <- "medv"
predictor <-"lstat"
ols_slr <- ols_estimators(data, predictor, respond)
boot_slr<-bootstrap_slr_summary(data, R = 1000,seed = NULL,predictor,respond)
plot_lr_bootstrap_scatter(ols_slr, boot_slr)
}
