% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_single_slr}
\alias{run_single_slr}
\title{Single simple linear regression with bootstrap}
\usage{
run_single_slr(
  data,
  respond = "medv",
  predictor = "lstat",
  R = 1000,
  seed = 2025,
  verbose = TRUE
)
}
\arguments{
\item{data}{data frame with Boston Housing data}

\item{respond}{response variable name (y). Cannot be "chas", "rad" or "b"}

\item{predictor}{predictor variable name (x)}

\item{R}{number of bootstrap replications}

\item{seed}{random seed for reproducibility}

\item{verbose}{print progress messages}
}
\value{
single_slr_result list
}
\description{
Check if the response variable is prohibited (excluding categorical variables
chas, rad, b), verify that the variable exists in the data, ensure that the
response and predictor variables are different, and use \verb{ols_estimators() to calculate the ordinary least squares correlation coefficient and perform R resampling through }bootstrap_slr_summary(). Integrate the results are and
output.
}
\examples{
data(BostonHousing)
data <- BostonHousing
run_single_slr(data, "medv", "lstat")
run_single_slr(data, respond = "rm", predictor = "ptratio", R = 2000)
}
