% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bootstrap_slr_summary}
\alias{bootstrap_slr_summary}
\title{Bootstrap analysis}
\usage{
bootstrap_slr_summary(data, R = 1000, seed = NULL, predictor, respond)
}
\arguments{
\item{data}{Boston Housoing dataset}

\item{R}{Number of bootstrap samples (set as 1000)}

\item{seed}{Input seed for reproducibility}

\item{predictor}{A string contains selecting variable name.}

\item{respond}{A string contains selecting variable name.}
}
\value{
A list containing: Bootstrap bias estimates,standard errors, etc
}
\description{
Bootstrap analysis
}
