% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate the correlation coefficient}
\usage{
calculate_correlation(data, predictor, respond)
}
\arguments{
\item{data}{A data frame containing the variables}

\item{predictor}{A string that contains the name of the predictor variable.}

\item{respond}{A string that contains the name of the response variable.}
}
\value{
Returns a list containing the followings:
correlation coefficient, p-value
}
\description{
Calculate the correlation coefficient between two variables and
perform a significance test.
}
\examples{
data <- BostonHousing
respond <- "medv"
predictor <-"lstat"
res <- calculate_correlation(data, predictor, respond)
print(res)
}
