% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_single_slr}
\alias{run_single_slr}
\title{Single simple linear regression with bootstrap}
\usage{
run_single_slr(
  data,
  respond = "medv",
  predictor = "lstat",
  R = 1000,
  seed = 2025,
  verbose = TRUE
)
}
\arguments{
\item{data}{data frame with Boston Housing data}

\item{respond}{response variable name (y). Cannot be "chas", "rad" or "b"}

\item{predictor}{predictor variable name (x)}

\item{R}{number of bootstrap replications}

\item{seed}{random seed for reproducibility}

\item{verbose}{print progress messages}
}
\value{
single_slr_result list
}
\description{
Single simple linear regression with bootstrap
}
\examples{
run_single_slr(data, "medv", "lstat")
run_single_slr(data, respond = "rm", predictor = "ptratio", R = 2000)
}
